# dev/apidoc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RESTful ToDo",
        "description": "Simple RESTful API for ToDo lists",
        "version": "1.0.0"
    },
    "paths": {
        "\/api\/todos": {
            "get": {
                "tags": [
                    "ToDo"
                ],
                "summary": "List all ToDo items",
                "operationId": null,
                "responses": {
                    "200": {
                        "description": "All existing ToDo items",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#\/components\/schemas\/ToDo"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorModel"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ToDo"
                ],
                "summary": "Add a new ToDo item",
                "operationId": null,
                "requestBody": {
                    "request": "ToDo",
                    "description": "The ToDo item to be created",
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/ToDo2"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The created ToDo item",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ToDo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/BadRequestError"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "\/api\/todos\/{id}": {
            "get": {
                "tags": [
                    "ToDo"
                ],
                "summary": "Show a single ToDo item",
                "operationId": null,
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ToDo item",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ToDo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ToDo item not found"
                    },
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorModel"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ToDo"
                ],
                "summary": "Update existing ToDo",
                "operationId": null,
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "requestBody": {
                    "request": "ToDo",
                    "description": "The ToDo item to be updated",
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/ToDo2"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The updated ToDo item",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ToDo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/BadRequestError"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorModel"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ToDo"
                ],
                "summary": "Delete a ToDo item",
                "operationId": null,
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ToDo deleted"
                    },
                    "404": {
                        "description": "ToDo item not found"
                    },
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "\/api\/doc.json": {
            "get": {
                "operationId": null,
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BadRequestError": {
                "description": "Bad request error model",
                "properties": {
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
            "ErrorModel": {
                "description": "General error model",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ToDo": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "tasks": {
                        "type": "array",
                        "items": {
                            "$ref": "#\/components\/schemas\/Task"
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ToDo2": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "tasks": {
                        "type": "array",
                        "items": {
                            "$ref": "#\/components\/schemas\/Task2"
                        }
                    }
                },
                "type": "object"
            },
            "Task": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Task2": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    }
}